StreetEdgeCarSpeedProvider is a 1st class citizen.
